bugfix: stop cursor jumping when typing quotes
We are using a function called normalizeQuotes in order to ensure
that people who type in a character like '’' will not be penalized
for their quote character.

The old way was to apply the normalizequotes function whenever a user
updated the text input field. However, this caused a subtle but where
if the user typed in the '’' character in the middle of the input text,
their cursor would jump all the way to the end

The fix was to let the user input their text normally (don't fix unusual
quote chars in the input field), but on submission of the field,
we call normalizeQuotes.