fix a terrible bug (saved wrong songlist)
Previously the code would compute the hash of the
full songlist, but then only save the serialized
version of the songlist that the player chose to
play the game with! This did not cause any issues
previously because, by pure luck, every time the
songlist was updated, the game would always be
first played with the full songlist. This time, I
chose to play the game with only 6 The Vault
tracks after updating the songlist, thus
triggering the bug because only 6 songs were saved
in the database instead of the full songlist.